backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
